Move CharacterCount to a new component
